+++ src/main/java/darwin/exception/DarwinException.java
@@ -1,0 +1,7 @@
+package darwin.exception;
+
+public class DarwinException extends Exception {
+    public DarwinException(String msg) {
+        super(msg);
+    }
+}
