+++ src/main/java/darwin/exception/IllegalTaskArgumentException.java
@@ -1,0 +1,7 @@
+package darwin.exception;
+
+public class IllegalTaskArgumentException extends TaskException {
+    public IllegalTaskArgumentException(String msg) {
+        super(msg);
+    }
+}
