+++ src/main/java/darwin/exception/IllegalTaskNumberException.java
@@ -1,0 +1,7 @@
+package darwin.exception;
+
+public class IllegalTaskNumberException extends TaskException {
+    public IllegalTaskNumberException(String msg) {
+        super(msg);
+    }
+}
