+++ src/main/java/darwin/exception/IllegalTaskTypeException.java
@@ -1,0 +1,7 @@
+package darwin.exception;
+
+public class IllegalTaskTypeException extends TaskException {
+    public IllegalTaskTypeException(String msg) {
+        super(msg);
+    }
+}
