+++ src/main/java/darwin/exception/TaskException.java
@@ -1,0 +1,7 @@
+package darwin.exception;
+
+public class TaskException extends DarwinException {
+    public TaskException(String msg) {
+        super(msg);
+    }
+}
