--- src/main/java/exception/DarwinException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class DarwinException extends Exception {
-    public DarwinException(String msg) {
-        super(msg);
-    }
-}
