--- src/main/java/exception/IllegalTaskArgumentException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class IllegalTaskArgumentException extends TaskException {
-    public IllegalTaskArgumentException(String msg) {
-        super(msg);
-    }
-}
