--- src/main/java/exception/IllegalTaskNumberException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class IllegalTaskNumberException extends TaskException {
-    public IllegalTaskNumberException(String msg) {
-        super(msg);
-    }
-}
