--- src/main/java/exception/IllegalTaskTypeException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class IllegalTaskTypeException extends TaskException {
-    public IllegalTaskTypeException(String msg) {
-        super(msg);
-    }
-}
