--- src/main/java/exception/TaskException.java
@@ -1,7 +1,0 @@
-package exception;
-
-public class TaskException extends DarwinException {
-    public TaskException(String msg) {
-        super(msg);
-    }
-}
