--- src/main/java/task/Event.java
@@ -1,34 +1,0 @@
-package task;
-
-import java.time.LocalDateTime;
-
-public class Event extends Task {
-    private String symbol = "E";
-    private LocalDateTime start;
-    private LocalDateTime end;
-
-    /**
-     * Initialises a task.Event task.Task with name, start and end.
-     * @param name A string of the task.Task's name.
-     * @param start A string indicating the task.Task's start date / time.
-     * @param end A string indicating the task.Task's end date / time.
-     */
-    public Event(String name, LocalDateTime start, LocalDateTime end) {
-        super(name);
-        this.start = start;
-        this.end = end;
-    }
-    @Override
-    public String getSymbol() {
-        return "E";
-    }
-
-    @Override
-    public String getTaskInfo() {
-        return super.getTaskInfo() + String.format(" (from: %s to: %s)", super.formatDate(this.start), super.formatDate(this.end));
-    }
-    @Override
-    public String toCsv() {
-        return super.toCsv() + "," + this.start + "," + this.end;
-    }
-}
